Stop missing frames!
Swap the policy for grabbing the latest image to drop fewer frames.

Previously, it would always pick the most recent image set for
processing. Now, it gathers a list of the last two hours of images,
and picks the first one newer than the last image processed.

The timestamp is initialized to early 2000, so the very first image
will be 5000 days or so newer than that bogus time.

This also includes significant cleanup encouraged by pylint. Some
renames, some rearranging.